Rewrite UTF-8 functions to actually be useful
Not to mention sane. The old functions were written in 2014 when I had a
*very* naive understanding of character encodings. They were written for
the sole purpose of base64-encoding Unicode in the browser, which wasn't
even implemented correctly then...

There are still some things that need fixing/adding; surrogate pairs and
options for stricter error-handling, which I'll probably do tomorrow.