Alphabetise named exports and test-suites
Something I should've done long ago, but never did because it would have
screwed up git-blame(1) and friends. This was bound to happen eventually
though, so it's best to get the diff-noise over with sooner than later.

Also added a script to ensure this shit doesn't need to be repeated; the
`sortn()` function will be moved to lib/misc.mjs in a follow-up commit.