Finish writing the majority of canvas specs
For now, at least. In addition to the issue described in 774dbcd, I need
to figure out why the `drawTextArea()` function isn't hyphenating input.
This is an old function written in mid/early-2014, and very visibly from
the ES5 era. It'll take a while to debug, so I'm leaving it for now.