Add function for identifying negative NaN values
This is admittedly useless, but I'm fascinated by the knowledge that NaN
values can hide information about their origin this way (IEEE-754 was an
eye-opening read, to say the least).

I have a recurring itch to write this function any time I see -0 and NaN
mixed together in JavaScript; Alhadis/Print@cf85002 was the final straw.